Modeled off of __torch_dispatch__, this PR demonstrates how to register
Python functions directly to the dispatcher.  Unlike __torch_dispatch__,
this makes it possible to directly override preexisting implementations
of PyTorch operators with alternate implementations, functionality that
maybe of interest to IPEX or Triton.

This is not a complete PR because it doesn't provide a coherent user
facing API for the functionality:

- We need to design an appropriate Python-side object hierarchy and
  API surface for interacting with the dispatcher.  Similarly, we need
  to battle harden the preexisting Python dispatcher API, which was
  previously solely used for testing.  Some things that need to be
  improved include file/line number reporting of registrations.

- For overriding use cases, we need to introduce new alias keys to
  allow for overriding without triggering a warning message that
  overriding is happening.

- When an override happens, we need a convenient way to call into the
  underlying implementation, in case we have decided we aren't
  interested in overriding the behavior anyway

- This PR is not compatible with torchdeploy; to handle torchdeploy,
  we must introduce a level of indirection whereby there is a single
  registered kernel for all Python interpreters, but it then determines
  whether or not to fallthrough to the base implementation based on
  whether or not the registration happened from a matching interpreter
  or not


Differential Revision: [D30074924](https://our.internmc.facebook.com/intern/diff/D30074924)

[ghstack-poisoned]

+++ torch/distributed/elastic/multiprocessing/api.py
@@ -14,7 +14,7 @@ import signal
 import subprocess
 import sys
 import time
-from contextlib import AbstractContextManager
+from contextlib import nullcontext
 from dataclasses import dataclass, field
 from enum import IntFlag
 from multiprocessing import synchronize
@@ -334,22 +334,9 @@ class PContext(abc.ABC):
             self._stderr_tail.stop()
 
 
-class _nullcontext(AbstractContextManager):
-    # TODO remove and replace in favor of contextlib.nullcontext
-    # when torch drops support for python3.6
-    def __init__(self, enter_result=None):
-        self.enter_result = enter_result
-
-    def __enter__(self):
-        return self.enter_result
-
-    def __exit__(self, *excinfo):
-        pass
-
-
 def get_std_cm(std_rd: str, redirect_fn):
     if IS_WINDOWS or IS_MACOS or not std_rd:
-        return _nullcontext()
+        return nullcontext()
     else:
         return redirect_fn(std_rd)
 
@@ -530,25 +517,26 @@ class MultiprocessContext(PContext):
     def _close(self, death_sig: signal.Signals, timeout: int = 30) -> None:
         if not self._pc:
             return
-        log.warning(f"Closing processes via signal {death_sig}")
         for proc in self._pc.processes:
-            try:
-                os.kill(proc.pid, death_sig)
-            except ProcessLookupError:
-                # If the process exited because of some reason,
-                # `ProcessLookupError` will be rasied, it is safe to ignore it.
-                pass
+            if proc.is_alive():
+                log.warning(f"Closing process {proc.pid} via signal {death_sig.name}")
+                try:
+                    os.kill(proc.pid, death_sig)
+                except ProcessLookupError:
+                    # If the process exited because of some reason,
+                    # `ProcessLookupError` will be rasied, it is safe to ignore it.
+                    pass
         end = time.monotonic() + timeout
         for proc in self._pc.processes:
             time_to_wait = end - time.monotonic()
             if time_to_wait <= 0:
                 break
             proc.join(time_to_wait)
-        log.warning(
-            f"Unable to shutdown processes via {death_sig}, forcefully exitting via {_get_kill_signal()}"
-        )
         for proc in self._pc.processes:
             if proc.is_alive():
+                log.warning(
+                    f"Unable to shutdown process {proc.pid} via {death_sig}, forcefully exitting via {_get_kill_signal()}"
+                )
                 try:
                     os.kill(proc.pid, _get_kill_signal())
                 except ProcessLookupError:
@@ -704,18 +692,27 @@ class SubprocessContext(PContext):
     def _close(self, death_sig: signal.Signals, timeout: int = 30) -> None:
         if not self.subprocess_handlers:
             return
-        log.warning(f"Sending processes {death_sig}")
         for handler in self.subprocess_handlers.values():
-            handler.close(death_sig=death_sig)
+            if handler.proc.poll() is None:
+                log.warning(
+                    f"Sending process {handler.proc.pid} closing signal {death_sig.name}"
+                )
+                handler.close(death_sig=death_sig)
         end = time.monotonic() + timeout
         for handler in self.subprocess_handlers.values():
             time_to_wait = end - time.monotonic()
             if time_to_wait <= 0:
                 break
-            handler.proc.wait(time_to_wait)
-        log.warning(
-            f"Unable to shutdown processes via {death_sig}, forcefully exitting via {_get_kill_signal()}"
-        )
+            try:
+                handler.proc.wait(time_to_wait)
+            except subprocess.TimeoutExpired:
+                # Ignore the timeout expired exception, since
+                # the child process will be forcefully terminated via SIGKILL
+                pass
         for handler in self.subprocess_handlers.values():
-            handler.close(death_sig=_get_kill_signal())
-            handler.proc.wait()
+            if handler.proc.poll() is None:
+                log.warning(
+                    f"Unable to shutdown process {handler.proc.pid} via {death_sig}, forcefully exitting via {_get_kill_signal()}"
+                )
+                handler.close(death_sig=_get_kill_signal())
+                handler.proc.wait()
